refactor(plot): rename PivotTablePlot -> PivotStatsPlot + unified table.plot.oncoplot() (S6)
The name PivotTablePlot wrongly implied it was the base for all PivotTable
plotting; it is really a set of one-shot statistical plots (PCA, annotated
boxplots). Rename to PivotStatsPlot and unify the entry point so the composite
oncoplot is reachable from the same accessor.

- git mv PivotTablePlot.py -> PivotStatsPlot.py; class renamed to PivotStatsPlot.
- PivotStatsPlot.oncoplot(**kwargs) returns an OncoPlot, so
  table.plot.oncoplot(figsize=...).main().add_freq().render() works alongside
  table.plot.plot_pca_samples(...).
- Backward compat: PivotTablePlot.py kept as a shim re-exporting PivotStatsPlot,
  so `from pymaftools.plot.PivotTablePlot import PivotTablePlot` still resolves.
  Not aliased at the package top level on purpose — the same-named shim submodule
  would shadow it (non-deterministic); canonical name is PivotStatsPlot.
- Updated ExpressionTablePlot base class, plot/__init__ exports, PivotTable.plot
  return type + docs.
- 2 new tests (unified entry, rename backward-compat).

Verified: 37 plot tests pass; StackingModel imports fine; eager path untouched.
